refactor: Use CSS Modules for PostPreview styles.
